core: Split file and improve module and macro name
