+++ relm4/src/safe_settings_and_actions/extensions.rs
@@ -4,367 +4,15 @@
  * SPDX-License-Identifier: (Apache-2.0 or MIT)
  */
 
-//! Utilities to safely manage actions and settings.
+//! Traits that extend types that use settings or actions to make them compatible with
+//! the types generated by [`safe_settings_or_actions!`](crate::safe_settings_or_actions).
 
-#![allow(unused_qualifications)]
-
+use super::*;
 use gio::prelude::{ActionExt, IsA, SettingsExt, SettingsExtManual, ToVariant};
 use gtk::{gio, glib};
 
-/// The safe extensions traits, without the safety traits.
-pub mod prelude {
-    pub use super::{
-        RelmMenu, SafeAction, SafeActionMap, SafeActionable, SafeApplication, SafeSettings,
-        SafeSimpleAction,
-    };
-
-    #[cfg(feature = "libadwaita")]
-    pub use super::SafeToast;
-}
-
-/// The basic trait for declaring setting safeties.
-pub trait SettingSafety {
-    /// The name of the setting or action.
-    ///
-    /// The macro [`safeties!`] automatically implements [`Display`](std::fmt::Display) using
-    /// this value, so it is not necessary to use it directly for loggin and similar cases.
-    const NAME: &'static str;
-
-    // NOTE: This method should be inside ActionSafety, but macro_rules! won't let me.
-    /// The `Variant` type concerning the [`Value`](WithValue::Value) of this safety.
-    fn variant_type() -> Option<std::borrow::Cow<'static, glib::VariantTy>>;
-}
-
-/// The basic trait for declaring action safeties.
-pub trait ActionSafety: SettingSafety {
-    /// The action name plus the group, as follows: `group.name`
-    const FULL_NAME: &'static str;
-
-    /// A self value.
-    const SELF: Self;
-}
-
-/// A trait that extends [`SettingSafety`] for safeties without value.
-///
-/// Any type that implements it must be [`DetailableSetting`] and must not implement [`WithValue`].
-pub trait WithoutValue: DetailableSetting {}
-
-/// A trait that extends [`SettingSafety`] for safeties with value.
-///
-/// Any type that implements it must not implement [`WithoutValue`].
-pub trait WithValue<'a>: SettingSafety {
-    /// Type used as parameter in safe methods.
-    type Value: glib::ToVariant + Into<glib::Variant>;
-
-    /// Type used to receive the value on signals.
-    type Mapping;
-
-    /// Function to convert [`glib::Variant`] to the [`Mapping`](WithValue::Mapping) type.
-    fn map(variant: &'a glib::Variant) -> Self::Mapping;
-}
-
-/// A trait that extends [`SettingSafety`] for safeties without state.
-///
-/// Any type that implements it must not implement [`Stateful`].
-pub trait Stateless: SettingSafety {}
-
-/// A trait that extends [`SettingSafety`] for safeties with state.
-///
-/// Any type that implements it must not implement [`Stateless`].
-pub trait Stateful<'a>: SettingSafety {
-    /// Type used as parameter in safe methods.
-    type State: glib::ToVariant;
-
-    /// Owned version of the [`State`](Stateful::State) type, used by getters or methods that mutate state.
-    type Owned: glib::ToVariant + glib::FromVariant;
-
-    /// Type used to receive the state on signals.
-    type Mapping;
-
-    /// Function to convert [`glib::Variant`] to the [`Mapping`](Stateful::Mapping) type.
-    fn map(variant: &'a glib::Variant) -> Self::Mapping;
-}
-
-/// Trait for setting or action safeties with value and without variants.
-///
-/// Any type that implements it must not implement [`DetailableSetting`].
-pub trait NotDetailable: SettingSafety {}
-
-/// Trait for setting safeties with variants, or without value and variants.
-///
-/// Detailable setting safeties facilitate settings maintenance.
-///
-/// Any type that implements it must not implement [`NotDetailable`].
-///
-/// ### Example:
-/// ~~~
-/// use relm4::gtk::prelude::{StaticVariantType, ToVariant};
-///
-/// relm4::safeties! {
-///     // Implements DetailableSetting:
-///     @state(param: bool)
-///     MySetting(name: "my-setting");
-///
-///     // Implements DetailableSetting:
-///     @value(param: &'a str, map: <str>)
-///     ColorScheme(name: "color-scheme") {
-///         System = ("System"),
-///         Light  = ("Light"),
-///         Dark   = ("Dark"),
-///     }
-///
-///     // Does not implement DetailableSetting:
-///     @value(param: i32)
-///     NotDetailable(name: "not-detailable");
-/// }
-/// ~~~
-/// See also: [`safeties!`], [`NotDetailable`].
-pub trait DetailableSetting: SettingSafety {
-    /// Function to convert [`glib::Variant`] to `Self`, especially when `Self` is an `enum`.
-    fn from_variant(variant: &glib::Variant) -> Self
-    where
-        Self: for<'a> WithValue<'a>;
-
-    /// Function to convert `Self` to [`glib::Variant`], especially when `Self` is an `enum`.
-    fn to_variant(&self) -> Option<glib::Variant>;
-}
-
-/// Trait for action safeties with variants, or without value and variants.
-///
-/// Detailable action safeties facilitate keyboard accelerators and settings maintenance.
-///
-/// Any type that implements it must not implement [`NotDetailable`].
-///
-/// ### Example:
-/// ~~~
-/// use relm4::gtk::prelude::{StaticVariantType, ToVariant};
-///
-/// relm4::safeties! {
-///     // Implements DetailableAction:
-///     MyAction(group: "win", name: "my-action");
-///
-///     // Implements DetailableAction:
-///     @state(param: bool)
-///     MySetting(group: "win", name: "my-setting");
-///
-///     // Implements DetailableAction:
-///     @value(param: &'a str, map: <str>)
-///     ColorScheme(group: "win", name: "color-scheme") {
-///         System = ("System"),
-///         Light  = ("Light"),
-///         Dark   = ("Dark"),
-///     }
-///
-///     // Does not implement DetailableAction:
-///     @value(param: i32)
-///     NotDetailable(group: "win", name: "not-detailable");
-/// }
-/// ~~~
-/// See also: [`safeties!`], [`NotDetailable`].
-pub trait DetailableAction: ActionSafety + DetailableSetting {
-    /// Gets the [detailed action name](https://docs.gtk.org/gio/type_func.Action.parse_detailed_name.html).
-    fn detailed_action_name(&self) -> &'static str;
-}
-
-#[macro_export]
-/// A macro to create unit structs or enums that implement [`SettingSafety`] / [`ActionSafety`],
-/// and depending on their characteristics also implements [`WithValue`] or [`WithoutValue`],
-/// [`Stateful`] or [`Stateless`], and [`DetailableSetting`] / [`DetailableAction`] or [`NotDetailable`].
-///
-/// Safeties can have a `@value`, which refers to the target value of
-/// some actionable methods or to settings values, as well as a `@state`.
-///
-/// `@value` and `@state` have the following parameters:
-/// 1. `param`: specifies the main data type and is used as a parameter of some safe methods. It is mandatory.
-/// 2. `owned`: an owned version of the param type. If not specified, the same as param is used.
-/// 3. `map`: a method of [`glib::Variant`] to convert it to the correct type data, used to receive the
-///           `@value` or `@state` in signals, or to retrieve an enumerated variant from [`glib::Variant`].
-///
-/// Unlike `param` and `owned`, `map` is specified as follows: `<method> ReturnType`.
-/// If not specified, it falls back to: `<get> OwnedType`.
-/// If only the method and not the return type is specified, falls back to: `<method> ParamType`.
-///
-/// In the case of `@value` it is not necessary `owned` if `map` is specified.
-///
-/// ### Syntax
-/// ~~~
-/// use relm4::gtk::prelude::{StaticVariantType, ToVariant};
-///
-/// relm4::safeties! {
-///     // Action safeties must specify the group parameter.
-///     UnitStruct(group: "action-group", name: "action-or-setting-name");
-///
-///     // If a lifetime is necessary, use 'a.
-///     @value(param: &'a str, owned: String, map: <str> &'a str)
-///     @state(param: &'a str, owned: String, map: <str> &'a str)
-///     Enum(name: "setting-name") {
-///         Variant = ("literal-value"), // ... Variants require @value.
-///         // In the case of string literals, they must always be enclosed
-///         // in parens for keyboard accelerators and settings to work.
-///     }
-/// }
-/// ~~~
-macro_rules! safeties {
-    (fallback! @($($foo:tt)+)            ) => { $($foo)+ };
-    (fallback! @($($foo:tt)+) $($bar:tt)+) => { $($bar)+ };
-
-    ( value:     $value:literal      ) => { $value };
-    ( value: (   $value:literal     )) => { $value };
-    ( value: [ $($value:tt),+ $(,)? ]) => { &[$($crate::safeties!(value: $value)),+] };
-    ( value: ( $($value:tt),+ $(,)? )) => {  ($($crate::safeties!(value: $value)),+) };
-    (detail:     $value:literal      ) => { $value };
-    (detail: (   $value:literal     )) => { concat!("'", $value, "'") };
-    (detail: [ $($value:tt),+ $(,)? ]) => { concat!("[" $(,$crate::safeties!(detail: $value),)','+ "]") };
-    (detail: ( $($value:tt),+ $(,)? )) => { concat!("(" $(,$crate::safeties!(detail: $value),)','+ ")") };
-
-    (setting: $name:literal;) => {
-        fn from_variant(_: &$crate::gtk::glib::Variant) -> Self { Self }
-        fn to_variant(&self) -> Option<$crate::gtk::glib::Variant> { None }
-    };
-    (action: $group:literal $name:literal;) => {
-        fn detailed_action_name(&self) -> &'static str {
-            <Self as $crate::safeties::ActionSafety>::FULL_NAME
-        }
-    };
-    (setting: $name:literal { $($variant:ident = $value:tt),+ $(,)? }) => {
-        fn from_variant(variant: &$crate::gtk::glib::Variant) -> Self {
-            match <Self as $crate::safeties::WithValue>::map(&variant) {
-                $($crate::safeties!(value: $value) => Self::$variant,)+
-                _ => unreachable!()
-            }
-        }
-        fn to_variant(&self) -> Option<$crate::gtk::glib::Variant> {
-            Some(match self {
-                $(Self::$variant => $crate::safeties!(value: $value).to_variant()),+
-            })
-        }
-    };
-    (action: $group:literal $name:literal { $($variant:ident = $value:tt),+ $(,)? }) => {
-        fn detailed_action_name(&self) -> &'static str {
-            match self {
-                $(Self::$variant => concat!($group, '.', $name, '(', $crate::safeties!(detail: $value), ')')),+
-            }
-        }
-    };
-
-    (safety!
-        $(#[$attr:meta])*
-        $(@value(
-            param:  $target_param:ty $(,
-            owned:  $target_owned:ty)? $(,
-              map: <$target_map:ident> $($target_mapping:ty)?)? $(,)?
-        ))?
-        $(@state(
-            param:  $state_param:ty $(,
-            owned:  $state_owned:ty)? $(,
-              map: <$state_map:ident> $state_mapping:ty)? $(,)?
-        ))?
-        $vis:vis $type:ident($(group: $group:literal,)? name: $name:literal $(,)?) $variants:tt
-    ) => {
-        $crate::safeties!(self_type! $(#[$attr])* $vis $type $variants);
-
-        impl $crate::safeties::SettingSafety for $type {
-            const NAME: &'static str = $name;
-            fn variant_type<'a>() -> Option<std::borrow::Cow<'static, $crate::gtk::glib::VariantTy>> {
-                $crate::safeties!(fallback! @(None) $(Some(<$target_param>::static_variant_type()))?)
-            }
-        }
-
-        impl std::fmt::Display for $type {
-            fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-                write!(f, "{:?}", <$type as $crate::safeties::SettingSafety>::NAME)
-            }
-        }
-
-        $(impl $crate::safeties::ActionSafety for $type {
-            const FULL_NAME: &'static str = concat!($group, '.', $name);
-            const SELF: Self = $crate::safeties!(self_constant! $type $variants);
-        })?
-
-        $crate::safeties! { fallback!
-            @(impl $crate::safeties::WithoutValue for $type { })
-            $(impl<'a> $crate::safeties::WithValue<'a> for $type {
-                type Value = $target_param;
-                type Mapping = $crate::safeties!(fallback! @(
-                    $crate::safeties!(fallback! @($target_param) $($target_owned)?)
-                ) $($($target_mapping)?)?);
-
-                $crate::safeties! { fallback!
-                    @(fn map(variant: &'a $crate::gtk::glib::Variant) -> $crate::safeties! {
-                        fallback! @($target_param) $($target_owned)?
-                    } { variant.get().unwrap() })
-
-                    $(fn map(variant: &'a $crate::gtk::glib::Variant) -> Self::Mapping {
-                        variant.$target_map().unwrap()
-                    })?
-                }
-            })?
-        }
-
-        $crate::safeties! { fallback!
-            @(impl $crate::safeties::Stateless for $type { })
-            $(impl<'a> $crate::safeties::Stateful<'a> for $type {
-                type State   = $state_param;
-                type Owned   = $crate::safeties!(fallback! @($state_param) $($state_owned)?);
-                type Mapping = $crate::safeties!(fallback! @(Self::Owned) $($state_mapping)?);
-
-                $crate::safeties! { fallback!
-                    @(fn map(variant: &'a $crate::gtk::glib::Variant) -> $crate::safeties! {
-                        fallback! @($state_param) $($state_owned)?
-                    } { variant.get().unwrap() })
-
-                    $(fn map(variant: &'a $crate::gtk::glib::Variant) -> Self::Mapping {
-                        variant.$state_map().unwrap()
-                    })?
-                }
-            })?
-        }
-
-        $crate::safeties! { detailable_or_not!
-            $type [$($target_param)?] [$variants]
-            impl $crate::safeties::DetailableSetting for $type {
-                $crate::safeties!(setting: $name $variants);
-            }
-            $(impl $crate::safeties::DetailableAction for $type {
-                $crate::safeties!(action: $group $name $variants);
-            })?
-            #[allow(dead_code)]
-            impl $type {
-                $crate::safeties!(value! $(
-                    $crate::safeties!(fallback! @($target_param) $($($target_mapping)?)?)
-                )? $variants);
-            }
-        }
-    };
-
-    (self_type! $(#[$attr:meta])* $vis:vis $type:ident;) => { $(#[$attr])* $vis struct $type; };
-    (self_type! $(#[$attr:meta])* $vis:vis $type:ident { $($variant:ident = $value:tt),* $(,)? }) => {
-        $(#[$attr])* $vis enum $type { $($variant),* }
-    };
-
-    (self_constant! $type:ident;) => { $type };
-    (self_constant! $type:ident { $variant:ident = $value:tt $(,$variants:ident = $values:tt)* $(,)? }) => { $type::$variant };
-
-    (detailable_or_not! $type:ty [$target:ty] [         ;  ] $($impl:tt)+) => { impl $crate::safeties::NotDetailable for $type { } };
-    (detailable_or_not! $type:ty [          ] [         ;  ] $($impl:tt)+) => { $($impl)+ };
-    (detailable_or_not! $type:ty [$target:ty] [$variants:tt] $($impl:tt)+) => { $($impl)+ };
-
-    (value! ;) => { };
-    (value! $target:ty { $($variant:ident = $value:tt),+ $(,)? }) => {
-        const fn value<'a>(&self) -> $target {
-            match self { $(Self::$variant => $crate::safeties!(value: $value)),+ }
-        }
-        const fn some<'a>(&self) -> Option<$target> {
-            Some(match self { $(Self::$variant => $crate::safeties!(value: $value)),+ })
-        }
-    };
-
-    ($($(#[$attr:meta])* $(@$keyword:tt $params:tt)* $vis:vis $type:ident $naming:tt $variants:tt)*) => {
-        $($crate::safeties!(safety! $(#[$attr])* $(@$keyword $params)* $vis $type $naming $variants);)*
-    };
-}
-
 /// Trait that extends [`gio::Action`] with action safety methods.
+#[allow(unused_qualifications)]
 pub trait SafeAction: gio::prelude::ActionExt {
     /// Safe version of [`state`](gio::prelude::ActionExt::state) for stateful action safeties.
     fn state_safe<'a, T: ActionSafety + Stateful<'a>>(&self, _safety: T) -> T::Owned {
